🌞 v0.5.0 - 시작 스크립트 작성

+++ src/main/kotlin/io/github/superjoy0502/socialdistancing/EventListener.kt
@@ -36,7 +36,7 @@ class EventListener : Listener {
             if (event.entity is Mob) {
                 val m: Mob = event.entity as Mob
                 if (m is Monster) {
-                    if (Random.nextFloat() <= 0.02) {
+                    if (Random.nextDouble() <= 0.02) {
                         virusMap.put(m.uniqueId, true)
 //                        logger.info("DEBUG: Mob with a virus has spawned")
                     }
